+++ test/models/agreement_event_tag_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class AgreementEventTagTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end